Testando rodar com scripts no maven

+++ dishfy/src/main/java/br/ufes/inf/dishfy/domain/Consumo.java
@@ -22,7 +22,7 @@ public class Consumo {
     private Date data;
 
     @Column(nullable = false)
-    float calorias; 
+    double calorias; 
 
     @ManyToOne
     private Receita receita;
